sunxi-disp: Set output-type and output-mode from the kernel cmdline
Because editing uEnv.txt is much simpler / more userfriendly
then dealing with script.bin, and human readable text strings are
way easier to remember then the hardwired mode table.

Note that if the cmdline arguments are not given, or not valid the
fex values are used as a fallback.

Note about the simple_strtol checkpatch warnings, we really want to use
simple_strtol here, as we need the ep to continue parsing after the number.

Signed-off-by: Hans de Goede <hdegoede@redhat.com>